refactor: remove annotation and change useEffect import
