+++ src/components/kakao-map.tsx
@@ -1,9 +1,9 @@
 'use client'
 
-import React from 'react'
+import { useEffect } from 'react'
 
 const KakaoMap = () => {
-  React.useEffect(() => {
+  useEffect(() => {
     window.kakao.maps.load(() => {
       var container = document.getElementById('map')
       var options = {
@@ -13,16 +13,11 @@ const KakaoMap = () => {
 
       var map = new window.kakao.maps.Map(container, options)
 
-      // 지도를 클릭한 위치에 표출할 마커입니다
       var marker = new window.kakao.maps.Marker({
-        // 지도 중심좌표에 마커를 생성합니다
         position: map.getCenter(),
       })
-      // 지도에 마커를 표시합니다
       marker.setMap(map)
 
-      // 지도에 클릭 이벤트를 등록합니다
-      // 지도를 클릭하면 마지막 파라미터로 넘어온 함수를 호출합니다
       window.kakao.maps.event.addListener(
         map,
         'click',
